feat!: collapse extraneous XBlock mixins
Various extraneous classes have been removed from the XBlock API, simplifying its implementation
and making debugging of XBlock instances easier. We believe that most, if not all, XBlock API users
will be unaffected by this change. Some improvements have also been made to the reference documentation.

See CHANGELOG.rst for details.

Closes: https://github.com/openedx/XBlock/issues/714
