+++ xblock/test/test_core.py
@@ -16,7 +16,7 @@ import pytest
 from opaque_keys.edx.locator import LibraryUsageLocatorV2, LibraryLocatorV2
 from webob import Response
 
-from xblock.core import XBlock
+from xblock.core import Blocklike, XBlock
 from xblock.exceptions import (
     XBlockSaveError,
     KeyValueMultiSaveError,
@@ -24,9 +24,8 @@ from xblock.exceptions import (
     DisallowedFileError,
     FieldDataDeprecationWarning,
 )
-from xblock.fields import Dict, Float, Integer, List, Set, Field, Scope, ScopeIds
+from xblock.fields import Dict, Float, Integer, List, Set, Field, Scope, ScopeIds, String
 from xblock.field_data import FieldData, DictFieldData
-from xblock.mixins import ScopedStorageMixin
 from xblock.runtime import Runtime
 
 from xblock.test.tools import (
@@ -386,8 +385,9 @@ def test_json_field_access():
             """Convert a datetime object to a string"""
             return value.strftime("%m/%d/%Y")
 
-    class FieldTester(ScopedStorageMixin):
-        """Toy class for ModelMetaclass and field access testing"""
+    @Blocklike.needs("field-data")
+    class FieldTester(Blocklike):
+        """Toy class for field access testing"""
 
         field_a = Date(scope=Scope.settings)
         field_b = Date(scope=Scope.content, default=datetime(2013, 4, 1))
@@ -435,8 +435,9 @@ def test_defaults_not_shared():
 
 def test_object_identity():
     # Check that values that are modified are what is returned
-    class FieldTester(ScopedStorageMixin):
-        """Toy class for ModelMetaclass and field access testing"""
+    @Blocklike.needs("field-data")
+    class FieldTester(Blocklike):
+        """Toy class for field access testing"""
         field_a = List(scope=Scope.settings)
 
     def mock_default(block, name):
@@ -475,8 +476,9 @@ def test_object_identity():
 
 def test_caching_is_per_instance():
     # Test that values cached for one instance do not appear on another
-    class FieldTester(ScopedStorageMixin):
-        """Toy class for ModelMetaclass and field access testing"""
+    @Blocklike.needs("field-data")
+    class FieldTester(Blocklike):
+        """Toy class for field access testing"""
         field_a = List(scope=Scope.settings)
 
     field_data = MagicMock(spec=FieldData)
@@ -777,17 +779,20 @@ def test_handle_shortcut():
 
 
 def test_services_decorators():
-    # A default XBlock has requested no services
-    xblock = XBlock(None, None, None)
-    assert not XBlock._services_requested
-    assert not xblock._services_requested
+
+    class NoServicesBlock(XBlock):
+        """XBlock requesting no services"""
+
+    no_services_block = NoServicesBlock(None, None, None)
+    assert not NoServicesBlock._services_requested
+    assert not no_services_block._services_requested
 
     @XBlock.needs("n")
     @XBlock.wants("w")
     class ServiceUsingBlock(XBlock):
         """XBlock using some services."""
 
-    service_using_block = ServiceUsingBlock(None, scope_ids=None)
+    service_using_block = ServiceUsingBlock(None, scope_ids=Mock())
     assert ServiceUsingBlock._services_requested == {
         'n': 'need', 'w': 'want'
     }
@@ -807,11 +812,12 @@ def test_services_decorators_with_inheritance():
     class SubServiceUsingBlock(ServiceUsingBlock):
         """Does this class properly inherit services from ServiceUsingBlock?"""
 
-    sub_service_using_block = SubServiceUsingBlock(None, scope_ids=None)
+    sub_service_using_block = SubServiceUsingBlock(None, scope_ids=Mock())
     assert sub_service_using_block.service_declaration("n1") == "need"
     assert sub_service_using_block.service_declaration("w1") == "want"
     assert sub_service_using_block.service_declaration("n2") == "need"
     assert sub_service_using_block.service_declaration("w2") == "want"
+    assert sub_service_using_block.service_declaration("field-data") == "need"
     assert sub_service_using_block.service_declaration("xx") is None
 
 
@@ -972,7 +978,7 @@ class OpenLocalResourceTest(unittest.TestCase):
         "public/\N{SNOWMAN}.js",
     )
     def test_open_good_local_resource(self, uri):
-        loadable = self.LoadableXBlock(None, scope_ids=None)
+        loadable = self.LoadableXBlock(None, scope_ids=Mock())
         with patch('pkg_resources.resource_stream', self.stub_resource_stream):
             assert loadable.open_local_resource(uri) == "!" + uri + "!"
             assert loadable.open_local_resource(uri.encode('utf-8')) == "!" + uri + "!"
@@ -986,7 +992,7 @@ class OpenLocalResourceTest(unittest.TestCase):
         "public/\N{SNOWMAN}.js".encode(),
     )
     def test_open_good_local_resource_binary(self, uri):
-        loadable = self.LoadableXBlock(None, scope_ids=None)
+        loadable = self.LoadableXBlock(None, scope_ids=Mock())
         with patch('pkg_resources.resource_stream', self.stub_resource_stream):
             assert loadable.open_local_resource(uri) == "!" + uri.decode('utf-8') + "!"
 
@@ -1000,7 +1006,7 @@ class OpenLocalResourceTest(unittest.TestCase):
         "static/\N{SNOWMAN}.js",
     )
     def test_open_bad_local_resource(self, uri):
-        loadable = self.LoadableXBlock(None, scope_ids=None)
+        loadable = self.LoadableXBlock(None, scope_ids=Mock())
         with patch('pkg_resources.resource_stream', self.stub_resource_stream):
             msg_pattern = ".*: %s" % re.escape(repr(uri))
             with pytest.raises(DisallowedFileError, match=msg_pattern):
@@ -1016,7 +1022,7 @@ class OpenLocalResourceTest(unittest.TestCase):
         "static/\N{SNOWMAN}.js".encode(),
     )
     def test_open_bad_local_resource_binary(self, uri):
-        loadable = self.LoadableXBlock(None, scope_ids=None)
+        loadable = self.LoadableXBlock(None, scope_ids=Mock())
         with patch('pkg_resources.resource_stream', self.stub_resource_stream):
             msg = ".*: %s" % re.escape(repr(uri.decode('utf-8')))
             with pytest.raises(DisallowedFileError, match=msg):
@@ -1038,7 +1044,7 @@ class OpenLocalResourceTest(unittest.TestCase):
         "static/\N{SNOWMAN}.js",
     )
     def test_open_local_resource_with_no_resources_dir(self, uri):
-        unloadable = self.UnloadableXBlock(None, scope_ids=None)
+        unloadable = self.UnloadableXBlock(None, scope_ids=Mock())
 
         with patch('pkg_resources.resource_stream', self.stub_resource_stream):
             msg = "not configured to serve local resources"
@@ -1078,6 +1084,12 @@ class TestIndexResults(unittest.TestCase):
         Class to test default Xblock provides a dictionary
         """
 
+    class TestXBlockWithDisplayName(XBlock):
+        """
+        Class to test default Xblock provides a dictionary
+        """
+        display_name = String(default="Test Block with Display Name", scope=Scope.settings)
+
     class TestIndexedXBlock(XBlock):
         """
         Class to test when an Xblock provides a dictionary
@@ -1089,13 +1101,24 @@ class TestIndexResults(unittest.TestCase):
                 "text_block": "Here is some text that was indexed",
             }
 
-    def test_default_index_view(self):
+    def test_default_index_view_with_no_display_name(self):
         test_runtime = TestRuntime(services={'field-data': DictFieldData({})})
         test_xblock = self.TestXBlock(test_runtime, scope_ids=Mock(spec=ScopeIds))
 
         index_info = test_xblock.index_dictionary()
         self.assertFalse(index_info)
         self.assertTrue(isinstance(index_info, dict))
+
+    def test_default_index_view_with_display_name(self):
+        test_runtime = TestRuntime(services={'field-data': DictFieldData({})})
+        test_xblock = self.TestXBlockWithDisplayName(test_runtime, scope_ids=Mock(spec=ScopeIds))
+        test_xblock.display_name = "My Display Name"
+
+        index_info = test_xblock.index_dictionary()
+        assert index_info == {
+            "content": {"display_name": "My Display Name"},
+            "content_type": "Test Block with Display Name",
+        }
 
     def test_override_index_view(self):
         test_runtime = TestRuntime(services={'field-data': DictFieldData({})})
